requirements.txt
alembic==1.14.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
dogpile.cache==1.3.3
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
keystoneauth1==5.9.1
Mako==1.3.7
MarkupSafe==3.0.2
marshmallow==3.23.2
mypy==1.14.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
openstacksdk==4.2.0
os-service-types==1.7.0
packaging==24.2
paramiko==3.5.0
pbr==6.1.0
platformdirs==4.3.6
psutil==6.1.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requestsexceptions==1.4.0
rq==2.1.0
SQLAlchemy==2.0.36
stevedore==5.4.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
